separate global_state macro related types
